Create 375.py
Print a new number by adding one to each of its digits

A number is an input in computer then a new no should get printed by adding one to each of its digits. 
If you encounter a 9, insert a 10 (don't carry over, just shift things around).

For example, 998 becomes 10109.

This challenge is trivial to do if you map it to a string to iterate over the input, operate, and then cast it back.
Instead, try doing it without casting it as a string at any point, keep it numeric (int, float if you need it) only.